fix errors caused by rails upgrade
The biggest issue was that active_record's serialize method
now seems to be making a lot more calls to load and dump, which
are not terribly efficient methods in CoughDrop. So I pulled
out serialize completely and implemented the parts that we actually
use.

The only breaking change was that it required calls to secure_serialize
to be made after has_paper_trail, mainly because secure_serialize
now overwrites the before_save method and manually calls any
additional before_saves (handling return falses correctly still) to
ensure that it has the last word when checking for changes to the
serialized column.